drivers: nvmem: Add add nvmem_fm25xxx driver
Zephyr has the eeprom_fm25xxx driver implemented
based on the EEPROM device driver API. Since we
introduced the more general nvmem device driver,
we implemented the nvmem_fm25xxx driver based
on the nvmem device driver API.
