+++ drivers/eeprom/Kconfig
@@ -99,7 +99,6 @@ source "drivers/eeprom/Kconfig.tmp11x"
 source "drivers/eeprom/Kconfig.xec"
 source "drivers/eeprom/Kconfig.mb85rcxx"
 source "drivers/eeprom/Kconfig.mb85rsxx"
-source "drivers/eeprom/Kconfig.fm25xxx"
 
 config EEPROM_SIMULATOR
 	bool "Simulated EEPROM driver"
