--- dts/bindings/mtd/infineon,fm25xxx.yaml
@@ -1,13 +1,0 @@
-# Copyright (c) 2025 Luna Pes <zephyr@orangemurker.com>
-# SPDX-License-Identifier: Apache-2.0
-
-description: Infineon FM25XXX SPI FRAM
-
-compatible: "infineon,fm25xxx"
-
-include: ["eeprom-base.yaml", spi-device.yaml]
-
-properties:
-  size:
-    required: true
-    description: Total FRAM size in bytes.
